services/requirements.txt
beautifulsoup4
flask-restplus
Flask-SQLAlchemy
google-cloud-storage
mysql-connector
nltk
numpy
pandas
pymysql
scikit-learn
tensorflow==2.0.0b1
traceback2
Werkzeug
